Comark.Syntax

Description

A definition of Commonmark's AST

Documentation

newtype Doc t

A Document

Constructors

Doc (Blocks t)

Block Elements

type Blocks t = Seq (Block t)

data Block t

Block elements

Constructors

ThematicBreak
Thematic break

Heading HeadingLevel (Inlines t)
Heading: level, sequnce of inlines that define content

CodeBlock (Maybe t) t
Block of code: info string, literal content

HtmlBlock t
Raw HTML Block

Para (Inlines t)
Paragraph (a grouped sequence of inlines)

Quote (Blocks t)
Block Quote (a quoted sequence of blocks)

List ListType Bool (Seq (Blocks t))
List: Type of the list, tightness, a sequnce of blocks (list item)

data HeadingLevel

Constructors

Heading1

Heading2

Heading3

Heading4

Heading5

Heading6

data ListType

Constructors

Ordered Delimiter Int

Bullet BulletMarker

data Delimiter

Constructors

Period

Paren

data BulletMarker

Constructors

Minus
-

Plus
+

Asterisk
*

Inline Elements

type Inlines t = Seq (Inline t)

data Inline t

Inline elements

Constructors

Str t
Text (string)

Code t
Inline code

Emph (Inlines t)
Emphasized text (a sequence of inlines)

Strong (Inlines t)
Strongly emphasized text (a sequence of inlines)

Link (Inlines t) t (Maybe t)
Hyperlink: visible link text (sequence of inlines), destination, title

Image (Inlines t) t (Maybe t)
Image hyperlink: image description, destination, title

RawHtml t
Inline Raw HTML tag

SoftBreak
A regular linebreak. A conforming renderer may render a soft line break in HTML either as line break or as a space.

HardBreak
A line break that is marked as hard (either with spaces or backslash, see the spec for details). In html it would be rendered as /

normalize :: Monoid t => Inlines t -> Inlines t

Consolidate adjacent text nodes

asText :: (Monoid a, IsString a) => Inline a -> a

Extract textual content from an inline. Note that it extracts only the primary content (the one that is shown in first place). For example it wouldn't extract an URL from the link.